scripts/alerting_function/requirements.txt
google-cloud-bigquery>=3.0.0
google-cloud-pubsub>=2.0.0
google-cloud-monitoring>=1.0.0
requests>=2.28.0
functions-framework>=3.0.0 